UMACONNECT/src/main/java/ES/UMA/UMACONNECT/UMACONNECT/SERVICE/CustomUserDetails.java
package ES.UMA.UMACONNECT.UMACONNECT.SERVICE;

import java.util.Collection;

import org.springframework.security.core.GrantedAuthority;
import org.springframework.security.core.userdetails.UserDetails;

import ES.UMA.UMACONNECT.UMACONNECT.ENTITY.User;

public class CustomUserDetails implements UserDetails{
	
	private static final long serialVersionUID = 1L;
	
	private User user;
	
	public CustomUserDetails(User user) {
		this.user = user;
	}

	@Override
	public Collection<? extends GrantedAuthority> getAuthorities() {
		return null;
	}

	@Override
	public String getPassword() {
		return user.getPassword();
	}

	@Override
	public String getUsername() {
		return user.getUsername();
	}
	
	public String getEmail() {
		return user.getEmail();
	}
	
	public String getGustos(){
		return user.getGustos();
	}

	@Override
	public boolean isAccountNonExpired() {
		return true;
	}

	@Override
	public boolean isAccountNonLocked() {
		return true;
	}

	@Override
	public boolean isCredentialsNonExpired() {
		return true;
	}

	public String getFullName() {
		return user.getNombre() + " " + user.getApellido1() + " " + user.getApellido2();
		}

	@Override
	public boolean isEnabled() {
		return true;
	}

}
